준수님_0526_템플릿,패키지 추가 pr

+++ src/main/java/com/team/berp/domain/Company.java
@@ -1,3 +1,4 @@
+
 package com.team.berp.domain;
 
 import jakarta.persistence.*;
@@ -9,7 +10,7 @@ import lombok.Data;
 public class Company {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer companyId;
+    private Long companyId;
 
     @Column(nullable = false, length = 100)
     private String companyName;
@@ -50,5 +51,6 @@ public class Company {
 
     @Column(nullable = false, length = 1)
     private String useYn = "Y";
+
 
 }
